src/requirements.txt
Django==4.2.14
django-admin-rangefilter==0.13.1
django-jazzmin==3.0.0
djangorestframework==3.15.2
psycopg2-binary==2.9.9
typing_extensions==4.12.2